Move live CallWithChat tests one folder in
